backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.12
httpx>=0.28.0
beautifulsoup4>=4.12.0
python-docx>=1.1.0
pdfplumber>=0.11.0
reportlab>=4.0.0
pdf2image>=1.17.0
openai>=1.57.0
langfuse>=2.53.0
stripe>=11.3.0
resend>=2.5.0
python-dotenv>=1.0.1
pypdf>=5.0.0
# Observability
sentry-sdk[fastapi]>=2.18.0
structlog>=24.4.0
# Rate limiting
slowapi>=0.1.9
